pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | migrations
    | frontend
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
known_django=["corsheaders","dotenv","django","drf_spectacular","rest_framework","slugify","httpx","aiogram","aiohttp","fluent","pandas","pydantic"]
known_thirdparty=["api","core","yandex_weather"]
sections=["FUTURE","STDLIB","DJANGO","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]

[tool.poetry]
name = "yandex_weather_bot"
version = "0.1.0"
description = "Bot for yandex weather"
authors = ["Labutin Andrew"]

[tool.poetry.dependencies]
python = "^3.11.5"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
drf-spectacular = "^0.26.2"
python-dotenv = "^1.0.0"
python-slugify = "^8.0.1"
django-cors-headers = "^4.0.0"
python-dateutil = "^2.8.2"
factory-boy = "^3.2.1"
httpx = "^0.26.0"
aiogram = "^3.0.0"
pandas = "^2.1.0"
pydantic-settings = "^2.0.3"
fluent-runtime = "^0.4.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
isort = "^5.9.3"
flake8 = "^6.0.0"
ipython = "^8.10.0"
django-extensions = "^3.1.3"
django-debug-toolbar = "4.1.0"
flake8-docstrings = "^1.7.0"
pytest-django = "^4.5.2"
pytest-pythonpath = "0.7.4"
pytest-sugar = "^0.9.4"
pytest-deadfixtures = "^2.2.1"
pytest-lazy-fixture = "^0.6.3"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
